added tests for new refactored code and endpoints have changed
