+++ api/googleauth.js
@@ -1,4 +1,5 @@
 // All Google auth code from their example AUTH stuff
+// TODO - Needs tests written for it
 let fs = require('fs');
 let readline = require('readline');
 let google = require('googleapis');
