+++ api/requests.js
@@ -12,7 +12,7 @@ module.exports = {
         calendarId: 'primary',
         timeMin: (new Date()).toISOString(),
         //   timeMax: (new Date()).toISOString(),
-        maxResults: 100 // to be edited later
+        maxResults: 1 // to be edited later
       }, function (err, response) {
         if (err) {
           console.log('ERROR:');
